+++ oshi-common/src/main/java/oshi/driver/common/windows/wmi/WmiResult.java
@@ -14,6 +14,36 @@ import org.jspecify.annotations.Nullable;
 public interface WmiResult<T extends Enum<T>> {
 
     /**
+     * Returns a result with no rows, for a query which was not executed.
+     *
+     * @param <T> the enum type representing the queried properties
+     * @return an empty result
+     */
+    static <T extends Enum<T>> WmiResult<T> empty() {
+        return new WmiResult<T>() {
+            @Override
+            public int getResultCount() {
+                return 0;
+            }
+
+            @Override
+            public @Nullable Object getValue(T property, int index) {
+                return null;
+            }
+
+            @Override
+            public int getVtType(T property) {
+                return WmiConstants.VT_EMPTY;
+            }
+
+            @Override
+            public int getCIMType(T property) {
+                return WmiConstants.CIM_EMPTY;
+            }
+        };
+    }
+
+    /**
      * Gets the number of results in this result set.
      *
      * @return the result count